Extract endless scrolling row from onboarding flow
